Get intial robot pose through topic instead of TF
The previous implementation retrieved the initial robot pose
through TF. However, for some unknown reason, it had a race condition
where TF provided some weird values for the very first transfom.
This should be investigated in the future and it might be due to
our implementation of the franka_state_controller where we publish
the TF for the endeffector. The new implementation retrieves the initial
robot pose thourgh the published FrankaState.