Fix linker errors when building with -DFranka_DIR
This problems occurred if building against a newer version of `libfranka` with
`-DFranka_DIR` while an older version of `ros-kinetic-libfranka` was still
installed.

Q: Why is `Franka_LIBRARIES` set to the target instead of an absolute path to
   `libfranka.so`?
A: If we link against the absolute path instead of Franka::Franka, the rpath
   isn't set while linking and the library can't be found at runtime. This goes
   away if we don't set `VERSION`/`SOVERSION` on `libfranka`, but we don't want
   to do that.

Q: Why do we also need `Franka_INCLUDE_DIRS`, as linking against the target
   sets up an include directory?
A: Linking against the target adds the include directory to the end, thus
   preferring (older) `libfranka` headers in `/opt/ros/kinetic`.

Q: Why do we need to add `SYSTEM` to `target_include_dirs`?
A: The include dirs for libfranka are added twice, once by linking to
   Franka::Franka (`-isystem`) and once by adding `Franka_INCLUDE_DIRS` (`-I`).
   However, CMake only adds the `-isystem` include, meaning that older
   `libfranka` headers from `-I/opt/ros/kinetic/include` (from
   `catkin_INCLUDE_DIRS`) are preferred. The solution is to include
   `/opt/ros/kinetic/include` with `-isystem`, too.