CHANGE: Define `ModelBase` abstract class
This interface is put into the `franka_hw::FrankaModelHandle` instead of the
plain `franka::Model` like previously. This allows for users to define their own
implementation of the type put into the Model Handle. This loosens the coupling
between the FrankaModelInterface & -Handle and the libfranka `franka::Model`
class.

Also we offer an alternative constructor to `franka_hw::FrankaModelHandle` which
takes an abstract base class pointer of the new `ModelBase` directly.