CHANGE: Lookup equilibrium_pose topic relatively
This is needed for nodes controller managers which spawn the controllers in
different namespaces. For example the model spawner of gazebo_ros_control puts
the controller names always beneath the model names, e.g.

/panda/cartesian_impedance_example_controller/equilibrum_pose

By not looking up the topic name absolutely (without leading '/') and a remap in
the launch file, the example still works but supports also the Gazebo naming