CHANGE: Augment EE + K Frame in `pose()` & `zeroJacobian()`
Unfortunately once a `KDL::Joint()` is created its `KDL::Frame` (pose) cannot be
altered anymore. This means, we have to add the two segments *each time the
pose/jacobian methods are called* to the chain and remove them afterwards again.
This is not the best regarding performance, but internally it seems that
KDL::Chain is just a std::vector of KDL::Segments, so its hopefully not awefully
slow.

Also since we are in a simulation, we have no hard realtime constraint in terms
of performance.