ADD: Hand URDF Macro with optional Gazebo Inertias
Note the deprecation warning in the `panda_arm_hand.urdf.xacro` because its
easier to if/else in XACRO then in a launch file. When you have a boolean arg in
a launch file you can't do anything other really than pass it onto XACRO. In
XACRO then you can use `<xacro:if value="$(arg ...)">`.

We keep the file, though, for backwards compatibility.