ADD: Proper Friction Parameters for fingers
These parameters make it so that the two fingers are stable and the object in
between does not bounce back and forth between the contact of the left and right
finger. This is accomplished by two parameters mainly:

* `max_vel = 0` inhibits any "snapping" motion, when the object and one finger
go beyond the penetration depth
* `min_depth` defines the allowed penetration depth

NOTE: That these parameter have to be set also on the object to grasp!