FIX: Use XYZ/RPY offsets in URDF not in Gazebo Spawner
Problem is, when you hand the parameters to the gazebo model spawner, the models
will appear correctly in gazebo at the specified pose, but Gazebo will publish a
TF from `world` -> `$(arm_id)_link0` with identity. This means that when placing
multiple pandas, they appear correctly in Gazebo, however ROS "thinks" they all
overlap and are attached to the world's origin at (0,0,0).

To fix this, we don't pass the pose to the model spawner, but rather to the URDF
directly. In here, the values specifiy the `<origin>` of the joint from `world`
-> `$(arm_id)_link0`, which then is published correctly as TF in ROS.