ADD: Implementation for Body Jacobian in `ModelKDL`
KDL offers a method to change the base of the jacobians, with which the velocity
mappings can be expressed in an arbitrary frame, not only the O frame.

Meaning to calculate the body jacobian out of the zero jacobian (relative to O
frame) we simple call this `KDL::Jacobian::changeBase()` function and give it
the rotation to change the basis.

This rotation comes from the FK of that frame, but has to be inverted for some
reason (not completely sure). The result of the calculations are checked vs the
model library with unit tests.

NOTE: There is also a `KDL::Jacobian::changeRefFrame()` function, which takes
also the position of the frame into account, which on first glance sounds more
fitting. However this does not match the behavior of our model library.