FIX: Double responding to interrupted actions
When an action B is called, while an action A is already running, the goal
callback of B will interrupt A. This will force a state transistion and make it
so that A will check its result and try to succeed or abort the action. However
since A was already aborted, you see the following error:

```
▏ERROR▕ /gazebo ⟩ ServerGoalHandle<ActionSpec>::setAborted() ⟩ L211 ⟫ To transition to an aborted state, the goal must be in a preempting or active state, it is currently in state: 4
```

State 4 from Actionlib is: `WAITING_FOR_CANCEL_ACK`

To fix the issue, we simple check in all action goal callbacks, after the state
transition happened, if the current goal is still active. If not, it was already
preempted already.