FIX: In HOLDING, let fingers track each other
When not doing this while an object is grasped the inerta of the fingers (and
the object itself) has an influence on the position of the fingers. This can
lead to the effect that both fingers drift inside the sledge of the gripper when
the EE is moved rapidly to the side.