Add 'ns' argument to allow specifying the robotNamespace for Gazebo
The robotNamespace is conceptually different from arm_id.
While the latter just acts as a name prefix for links and joints,
the <robotNamespace> tag specifies the name of the ROS namespace, which is used
for key robot components like the joint_states topic, the controller_manager, etc.

To allow for an empty namespace, the default needs to be specified as empty.
Otherwise, i.e. having a non-empty default, specifying ns:="" on the commandline
will fall back to the default despite an (empty) value was provided.