Allow starting the Gazebo robot w/ and w/o a namespace.
By default, we don't start in a namespace to match the real-robot behavior.
Unfortunately, we cannot just specify an empty ns="" in roslaunch, but we
need to distinguish the two cases empty/non-empty namespace.
To avoid code duplication, the relevant code was factored out into panda_ns.launch.xml,
which is include w/ or w/o a namespace.
The file-ending is deliberatively choosen as .launch.xml to hide the file from bash completion.