reliably launch joint_{position,velocity}_example_controller on startup in gazebo.
Gazebo initializes the joint positions **after** the spawning the model. Therefore, the robot is in zero position for a very short time. Since, the position and velocity example controller are checking the position first before starting, the controller failed ~50% of the time when starting it from the same launch file as the controller. Delaying the start of the controller spawner by 5 seconds fixes this