ADD: `<xacro:collision_capsule>` macro
This macro abstracts the creation of a "collision capsule" from two collision
spheres and a collision cylinder.

Since arbitraty RPY rotations are hard to calculate in XACRO (would require
rotation matrix multiplication and probably reduce redability) for now we only
allow the discrete 6 orientations which are aligned with X/Y/Z axes.