CHANGE: Restart all running controllers on error_recovery
This is similar to how `franka_control` does it. It it necessary since
we (franka_control & franka_hw_sim) keep the controllers running even if the
user stop is pressed in order for "read-only" controllers (such as
FrankaStateController) to keep updating. However, when the user stop error was
recovered (i.e. the controller's commands are fed through again) their will be a
large jump in desired vs actual control signals.

This is mitigated by restarting all controllers via the `switch_controller`
service call of the controller manager.