ADD: Make interactive marker configurable in `panda.launch`
In some scenarios we want to start the cartesian_impedance_example controller
but command the equilibrium pose by ourselves. A running `interactive_marker`
node will keep publishing onto the topic, making it compete with our publisher.

To avoid breaking changes we make the presence of the node also configurable. By
default it will still be started whenever the cartesian impedance controller is
started but this can be deliberately overwritten by the user.