fix: `error_recovery` action should not recover future errors (#316)
Previously, once `error_recovery` action is called during a motion, the action callback
function will wait at the "mutex lock" until the motion stops. If the motion
stops due to an error, the error will be recovered immediately which often
results in an unexpected behavior of the robot.

This fix lets the action callback return early if there is no error present.